--- TrackYouFy/app/src/main/res/layout/tab_layout.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:layout_height="wrap_content">
-
-
-    <android.support.design.widget.TabLayout
-        android:id="@+id/tabs"
-        app:tabGravity="fill"
-        app:tabMode="fixed"
-        android:background="@android:color/holo_blue_bright"
-        app:tabIndicatorColor="@android:color/holo_red_dark"
-        app:tabSelectedTextColor="@android:color/holo_green_dark"
-        app:tabTextColor="@android:color/white"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-    </android.support.design.widget.TabLayout>
-
-    <android.support.v4.view.ViewPager
-        android:background="@android:color/darker_gray"
-        android:id="@+id/viewpager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        >
-
-    </android.support.v4.view.ViewPager>
-
-
-
-</LinearLayout>
