+++ TrackYouFy/app/src/main/res/menu/drawer_menu.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
+    <group android:checkableBehavior="single">
     <item android:title="SPOTIFY"
         android:id="@+id/nav_item_spotify"
         android:icon="@drawable/spotify_icon"/>
@@ -12,5 +13,6 @@
     <item android:title="Podcast"
         android:id="@+id/nav_item_podcast"
         android:icon="@drawable/podcast_icon"/>
+    </group>
 
 </menu>
